chore: 🤖 refactor new conversation flow related code for origin
