+++ apps/web/src/app/(formSubmissionPage)/view/[formId]/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from "next";
 import { notFound } from "next/navigation";
 
-import { FormViewer } from "@/components/newFormViewer/formViewer";
+import { FormViewer } from "@/components/formViewer/formViewer";
 import { formUpdateSchema } from "@/lib/validations/form";
 import { api } from "@/trpc/server";
 import { FormSubmissionPageHeader } from "./_components/header";
