+++ packages/react/src/index.ts
@@ -1,3 +1,3 @@
-import { useNewConvoForm } from "./useNewConvoForm";
+import { useConvoForm } from "./useConvoForm";
 
-export { useNewConvoForm };
+export { useConvoForm };
